requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
openai>=1.50.0
anthropic>=0.40.0
playwright>=1.50.0
scrapy>=2.12.0
serpapi>=0.1.5
pinecone-client>=5.0.0
weaviate-client>=4.0.0
fastapi>=0.115.0
flask>=3.1.0
psycopg2-binary>=2.9.10
pymongo>=4.10.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.10.0
requests>=2.32.0
beautifulsoup4>=4.12.0
